warn about missing font files via printf, not debugmsg
debugmsg triggers a new ImGui frame, which messes up the font
atlas. printf is suboptimal, but I plan to introduce an in–game font
picker that can inform the user that the font file couldn’t be found
in a less heavy–handed and totalizing way than debugmsg. Therefore a
suppose printf will do for now.