Bionics no longer block conflicting mutations
Instead, the mutation can fail, inflicting some internal damage, or
destroy the bionic, causing large internal damage.

Most of the damage is through bleeding and pain, unless it's a
particularly severe injury, in which case limb damage is applied.

The exact numbers and chance in particular, don't have a particularly
strong backing.