Reduce price on vehicle to something a little more "gamey"
100k made a lot more sense to me. But at the price it would never make any sense to buy it over just buying parts and scraping a vehicle together yourslef or just trying to find on in Salus.

Also there just aren't vehicles spawning on Salus. So I almost want the price to be more reachable for the player given the current state of the mod. So people can play with vehicles sooner. I will update this price once we have more vehicles spawning and we know what it's worth.