Allow generic-factory loaded objects to defer
Mapgen updates used by mission defs decide to defer loading when they
run into something that is not defined. This is not a good pattern (they
should just load and then later be checked), but it's too deep to
untangle easily.

Allow instead those mission definitions request that generic factory
defer them, instead of hand-rolling their own deferral.

I think this is hack, but I don't have time to untangle the deeper mess
and I think it's worth fixing the issue.

This causes any issue in a mapgen update in a mission def to result in
the half-formed mission def being inserted into the factory, and then
the deferred JSON trying to insert it again, triggering the 'duplicate
object from same source' check.