Update EOCs.json with Vitrification and Incorporeal
If you fail the black glass house encounter, you are supposed to die and turn to glass. But if you play sky island, it saves you from said fate, and brings you to the island. Once there the black glass curse does not go away and you repeatedly "die" to the curse.

Adding vitrification and Incorporeal to the u_lose_effect, the player is no longer turning to glass or incorporeal after returning to the sky island.