Bugfix: Mi-go pathfinding avoids fire
Bugfix for pathfinding to not use the straight-line optimization when
there is some sort of non-normal tile between the source and target.
Mi-go will now path around fires if there is such a route.

There is an optimization in place before the A* pathfinder is invoked,
that checks for whether the target is reachable by a straight-line. In
such case, the A* pathfinder is not invoked, presumably because of
performance reasons. The evaluation of this straight-line optimization
does not use the same way of choosing whether or not each tile is
walkable or not. The real A* pathfinder will usually route around tiles
that are technically walkable but might be questionable (such as fires),
but the straight-line optimization only checks for whether it's
technically walkable. The real A* pathfinder might also pick a route
through such questionable tiles if there are no other choices. For
example, a monster might walk through fire if surrounded. So the
question of whether or not to pick the straight-line optimization is a
bit trickier than it might seem at first.

Because of these differences in the straight-line evaluation and the
real A* pathfinder, the straight-line optimization was previously chosen
even if the A* pathfinder would have picked a different path. For
example: before, a mi-go picked a straight-line path through fire, even
if it had `avoid_dangerous_fields` because, well, the tiles on fire are
technically walkable, right?

This commit instead aborts the straight-line optimization if any of the
tiles on the straight-line has any sort of non-normal characteristic. If
so, the A* pathfinder is invoked, which will choose the best route given
the circumstances. The route will probably walk around the fires, but if
there's no other way, it will walk through it.

This commit also adds `avoid_dangerous_fields` to all mi-gos, because
they're fairly intelligent and would likely not willingly walk into a
fire. There are likely more monsters that should also have the
`avoid_dangerous_fields`-flag set, such as most mammals?