Steel update: motors
The steel used in electric motors is chosen for its magnetic properties
and is a mild steel , or rather a specialty steel with low carbon
content (https://en.wikipedia.org/wiki/Electrical_steel).
Mild steel is a simplification, but should be a decent one.

I also added copper to the motor materials, for obvious reasons.