All in one can (EOC)
Condensed everything into a single EOC that creates the emit and removes it after a 1 sec delay with a new tr_furn_transform object.

With that I was able to use the same field_id that is attached to insecticide hardcode. Removed new, now redundant field_id.

That old field_id was updated to apply the new effects and the assoc. immunity data.

Adjusted duration, Str penalty, and cough chance of effects.

Future commit will include intensity levels for the insecticide in eyes and mouth effects.