[Bombastic Perks] Repeatable Stat Increases
Way back when I first saw Bombastic Perks, my first thought was that it could basically take the place of Stats Through Kills, only moreso. It does the same thing, letting you level up your character by killing things, but also has many more things to spend level ups on, and also has a lot more customizability in having a user-definable xp curve instead of a hardcoded list of xp thresholds. The only thing holding it back from that potential was that the base stat increases were pretty limited.

This is my attempt to add that functionality.

Basically what I've done is to add 4 dummy perks whose only function is to increment a variable and then delete themselves (so that they can later be added again). I then take those four variables, and add them as stat enhancements to the perk menu trait (the one trait that every user of the mod is guarenteed to have).

Alternatives:

My original hope was to be able to directly increase the base stat. However, I couldn't find a way to do that within EOCs (there may be one, this was kind of a learning project for me to try and figure out EOCs).

I did consider making a separate perk to hold the bonuses. However, the character sheet already gets crowded when using Bombastic Perks, and it doesn't do anything visible to the player anyway. Maybe there's a reason I'm not seeing why this is a bad idea? IDK.

I did of course consider making the perk selection call the EOC to increment the bonus variable directly instead of using dummy perks, but that would involve messing with the functionality of the perk menu itself, and ultimately that's a lot messier than just doing the dummy perk thing. The mod is designed around the idea that the only thing it's going to do to the player is add traits, I'll work with that instead of against it.

I tried a couple different ways to make the dummy perks activate an EOC then delete themselves. Having it just start already activated doesn't call the EOC correctly. I didn't want to make the player manually activate it every time, that's annoying and unnecessary. The "trigger" thing works fine, but it was surprisingly hard to find a way to create an always true conditional. Trigger wants a condition, conditions by their nature are designed to be conditional. Telling it to roll a 1 in 1 chance is kind of silly, but hey it works. There probably is a better way to do that; it seems like an oversight if there's not a better way to do it.

Testing:

I created a character, got some xp, and bought each perk at least once, and a couple of them multiple times. It does require closing and reopening the menu so that the perk has time to process itself, which shouldn't be a problem in normal play.