Atlatl fix
Atlatl reqs needed both fine cutting and a plank, but this is a weapon made in history before either of these things were around, so I changed some things. The recipe requires 2 sticks, as on of them would be shaped into a hook, and the other made into the spear thrower shape. In real life, you could do it with only a stick, but it'd need to either use a different method than shaft and hook, or would need to be a perfect stick (wishbone shaped).