Faction food supply can expire over time
Convert faction food supply tracking from a single set of nutrients to a
list of pairs of time_point, nutrients - this is the nutrients, and what
time they expire that. When evaluating food supply, discard any
nutrients that are expired. non-perishable nutrients use the special
case of calendar::turn_zero.

Food is consumed first from the soonest to expire entry, and only
consumed from the non-perishable stockpile once all perishable entries
have been consumed.

For simplicity, rot points are truncated to the most recent time point
that is zero mod 1_days (the nearest day), and all added foods with that
rot point are counted together.

Convert basecamp::camp_food_supply to simply supply calories from store,
addition must be done through the faction or by distributing through the
camp.

The impact of the "pantry" basecamp feature on food supply is no longer
counted, because rot is no longer being used to reduce calories added to
a permanent non-perishable store.

Dialogue can no longer modify faction camp food or vitamin stores
because there are no use cases. Because of the higher degree of
complication now, it would be better to be implement once a use case
emerges.

Faction templates can be defined with the full features, but it's very
opaque. This could be improved in the future.