Create a "thin" archive instead of copying objects
'ar' has the option to create a "thin" archive, which contains
"references to the original copies of the member files of the archive"
As these files will not move while building the tests, this is safe to
use, and significantly speeds up the archive generation.

On my system, this turns 10-13s building the archive into 0.7s

OSX, at least in our CI, does not use GNU ar and thus cannot do this.