Make monster attack cooldowns mandatory
Monster attacks defined in the `special_attacks` section of an mtype
definition are meant to inherit from `"type": "monster_attack"` entries
with the same id.

This is working, however the cooldown loading was not respecting the
`was_loaded` value, and thus would overwrite each time, making all
`special_attacks` that inherited and did not set a specific cooldown
reset the cooldown to 0.

Make the cooldown value mandatory, meaning that the first definition of
any monster attack must define a cooldown, but anything that inherits
(and thus was_loaded is set to true) will keep that value unless it
specifies a new one.

Also adds a regression test.