Fix mutation endless loop
Mutating generates a weighted_list of possible categories to mutate
into, then picks one, and removes it from the list. Then, it tries to
find a mutation in that category to mutate. If it fails, it tries
another category until the list is empty.

It used add_or_replace with the value of zero to remove the category -
replacing the entry with a weight of zero, meaning it could not be
selected.

In a757b516e51109284eaa9da4f9b2361434acb540 the condition of
add_or_replace was changed to reject all weights >= zero, instead of all
weights > 0, and as such the list of categories was never cleared,
meaning mutation attempts would continue forever if no mutations were
valid.

When a call was made to mutate() without enough primer for any of
the mutations, it would trigger this condition and loop forever.

Make the removal of the category explicit by using the `remove`
function, instead of this special behavior of add_or replace. Make
`add_or_replace` actually remove on weight 0, instead of ignoring the
call.