Only spawn monster if not already spawned
The test only needs a monster to be there, not a specific monster.
Because we're doing several iterations, the monster may already be
spawned in that spot from a previous iteration, and so don't try to
spawn it (and fail the test) if it already exists.

Also, check the terrain is set correctly _after_ the setup, in case the
setup does something (like clearing the map) that invalidates the test.