use was_loaded properly for bash and deconstruct
was_loaded indicates whether or not it is the first time this object has
been loaded.

The terrain/furniture loading code was passing the terrain/furniture
was_loaded value, when they could be creating the object and thus it
would be it's first time being loaded.

Pass whether or not the object has been created as the was_loaded value
instead of whether or not the terrain/furniture it's part of has.