Create dust when smashing furniture and terrain (#81565)
* Allow fields to choose to be moppable

Currently, all liquid fields are moppable. Allow fields that are not
liquid to select that they can also be cleaned with a mop.

* use was_loaded properly for bash and deconstruct

was_loaded indicates whether or not it is the first time this object has
been loaded.

The terrain/furniture loading code was passing the terrain/furniture
was_loaded value, when they could be creating the object and thus it
would be it's first time being loaded.

Pass whether or not the object has been created as the was_loaded value
instead of whether or not the terrain/furniture it's part of has.

* Allow bashed terrain/furniture to create fields

Allow specifying a field and an intensity to spawn that field at when an
object is hit, and when an object is destroyed. The field spawn when it
is hit will also spawn when it is destroyed.

* Add dust and splinters fields for bashed ter/furn

The dust is in the gas phase and can spread - this allows bashing to
spread dust over the place, instead of only in the tile bashed.

Similarly for splinters, but it's less likely to spread, and will try to
spread dust.

* Add dust to everything that splinters

All terrain/furniture that produce "splinter" in their bash results have
been given the dust and splinters field produced on bash.

* Mods furn/terrain smashes into dust

All terrain/furniture in mods that break into splinters produce the dust
and splinters field on bash.