Update Hair Dye Kit
Hair Dye Kit can now hold its hair dye since it has a box. I still need to go in and change the spawning to use the kit as a container and simply pick a dye to be in it. Also need to spawn them empty and add them to more locations