Shift test shooter farther from edge of map
There was an occasional segfault on the OSX test build due to a
projectile going out of bounds on the map, and likely causing some
unspecified undefined behavior.

The segfault has been fixed, but it now triggers a debugmsg which will
fail tests.

The culprit is likely that the test NPC who is firing a shotgun shoots a
projectile that due to dispersion is sufficiently knocked off course
that it crosses the 10 tiles from them to the map edge, and goes out of
bounds. Because there are multiple trials, and they are shooting a
shotgun with multiple projectiles, potentially after the object that
they are aiming for has been destroyed, it manages somewhat consistently
to trigger this unlikely case.