Case Hardening 2.0
Hardened steel katana: 2 steel, 20 hours, even discounting the work into making the steel a proper blade, most of the time would still be spent just working the steel into the process of case hardening.

2 steel, x 10, in 20 hours, leaves us with 20 steel for 20 hours, at 1 hour of work per session that still results into 1 steel, I left it at 4 to account for the time in the original that you would spend making the steel into a proper blade, and to contemplate a lot, but I mean a whole lot, of steel craking and the like when you start in your first tries, the carbon is also already at x10 of what the original spends, and in 1 hour I seriously doubt you could make use of more.

This recipe was already spending a lot of material, I should rather be decreasing it...

One must take into account in these cases that the steel used is actually 4 steel chunks per steel_standard, and the carbon is 25 units per carbon requirement (So 250 carbon for one hour of work, a lot, but the amount and time of carbon should be the most important aspects of learning to work with case hardening).