Quenching & Tempering 2.0
Tempered steel katana: It uses the same 2 steel, but takes even longer at 24 hours, same rationale as the last one.

Same as the last one, it uses 2 steel, for a 24 hour work, x10 is 20 steel, divided into 24 leave us with 0.8, even less than the last one, I leaved it at 4 too, there is just not that much time and what you need to learn is just the process of how much heat, for how much time, and how it changes the proprieties of the steel when working with this particular process.

Reduced the skill limit too.