Use damage/hp for bashed objects
Add functions for bash data to generate damage/hp values, and use those
instead of directly querying str_{min,max}. This allows expanding how
furniture responds to damage in the future.

The bool arguments could be improved, but work well enough.

Remove the bashing debugmode message for now, it can be re-enabled
later.

Sound shouldn't care about the supported/blocked strength, so just
ignore that, it won't be that different.