Character/map bash respond to all damage types
Pass in a map of damage types and the amount, instead of just a strength
number.

Add a member to damage types that allows specifying how they convert
into bash damage. Right now, this is zero for all types but bash.

There are rough edges right now where this interacts with pathfinding
and monster bashing: those have a strength integer that they just
convert to {bash: strength}

For simplicity, keep a map::bash overload that takes integer strength,
and just converts it bash damage for the real bash function.

Special mention to the {{{}}} constructor, which is
pair{} -> initializer_list{pair{}} -> map{initializer_list{pair{}}}