Backport molotov fixes to 0.I (#81848)
* Fix molotovs processing too many times

* More correct return value handling for iuse

iuse functions return an optional<int>, where it is nullopt if the item
is not used, and the contained value is either how many charges were
used or if it wants to be destroyed - depending on the context.

Make the second behavior explicit with a flag, and apply it to the
molotov which depends on the destruction behavior.

* Remove debugmsg for stable

This makes it so behavior is preserved _unless_ the flag is applied,
which is probably a fairly safe merge for stable. It's the same behavior
except in the known-buggy case.

---------

Co-authored-by: RenechCDDA <84619419+RenechCDDA@users.noreply.github.com>