Add bathroom hair dye spawn
Aight, everything appears to be all wrapped up now. I'm not adding the pocket to the hair dye kit and just spawn the dye with it. Not a lot of houses use the SUS group I modified, so I had to check several houses to find a bottle of dye.