upgrade to ImGui tag v1.91.9b
Our source layout is a different from ImGui’s (we take a subset of the
files, and we don’t split them into subdirectories). The new files
were simply copied in over the old. With this approach there can be no
merge conflicts, but any changes we’ve made to ImGui were obviously
overwritten.

I then examined the diff and reverted any changes that did not appear
to be related to the new version of ImGui. As most of those changes
came from IMTUI they were guarded by an `#ifdef IMTUI` already and
were very obvious. @ZhilkinSerg was kind enough to point out a few
important changes that I missed so I went back and recovered them. As
these changes were not related to IMTUI I added a comment to all of
them saying that we would want to keep the change in the future. This
should make further merges easier.