upgrade to ImGui tag v1.91.9b
Our source layout is a different from ImGui’s (we take a subset of the
files, and it splits them into more subdirectories). The new files
were copied in and merge the merge examined. With this approach there
can be no merge conflicts, but changes related to ImTui were kept as
were some other minor modifications that we have made to include paths
and some pragmas and other little things.

As all of the ImTui changes are guarded by an #ifdef IMTUI, they were
not hard to find. Most were kept, but two were discarded as surplus to
requirements. The compiler reminded me of the include path changes.