JSON: Report when delete doesn't delete
Unfortunately the only way to add context, which is necessary to
resolve this, is to throw a JsonError and abort loading.

Also have a debugmsg that reports the specific value when it is not
deleted, before the error. Add a bunch of template code to convert the
data into a string in any way it can.

This isn't reported for the bitsets, and it could, but I think it's
lower priority.