Small generic factory fixes + script to load game with all mods (#82283)
* Warn on disabled extended feature for mandatory

Some of these could definitely be supported.

* fix type for dbl_or_var reader

It doesn't matter because it overrides operator(), but the templated
type for generic_typed_reader should be the derived reader.

* Add script to load all mods

This allows easily checking all mod combinations for JSON errors

* Remove duplicate tests/cata_test check

Remove it and use the one below, which allows using the test binary from
different locations.