Update cranberry juice fun value and description.
I believe this better would reflect the value of cranberry juice, which is pretty sugary. Given that water is 0, and cranberry soda is 6 fun, this seems like a more realistic fun value, rather than -1. which you would think is the fun value for something unappetizing. Of course, this is subjective, but again, given that crispy cran is 6 fun, which seems to have a similar favor profile (Cranberry juice and lemon-lime soda), I think it's justified.