Special attacks to optional
This requires a really gross reader because there's two interrelated
bits of data spread across two different variables, and the data
definition is complex, so needs some special handling for delete.

The loading of the array syntax is gross too.

But the bugs being fixed shows it is an improvement :)