Prevent segfault when cancelling completed butchery
Prevents segfault when cancelling a butchery that has been completed on
the same turn.

According to the documentation of `butchery_activity_actor::bd`, when
the list is empty, we are done. The previous segfault happened because
we took `bd.back()` and used that object, but the list of butcheries was
empty. This might happen if, for example, we cancelled the activity
on the same turn that the last butchery was `pop_back()`-ed - e.g. the
activity was to be finished on the same turn that we cancelled.