Prevent segfault from teleport trap
Prevents segfault when a monster steps on a teleport trap and gets
teleported outside the map bubble.

When a monster gets teleported outside the map bubble, a temporary map
is created for the destination, in `teleport::teleport_to_point`.
Previously, the reference to this temporary map got lost in the lambda
being modified by this commit, since the lambda was `static` and thus
retained the reference to the map from where it was first invoked.