Revert "Add batch multiplier to crafting UI"
This reverts commit 86cd2e789ec19aed7255e3db0185cfbfedfb7494.

Crafting models a person creating things in a workshop, and batch time
scalings represent being able to amortize setup costs, or being able to
process multiple things at once (because your pot is large enough, etc).
This assumption breaks at large batch sizes (50 is already generous),
and processes which are legitimately batch-able at high rates deserve
simulation of that process, instead of extending the entire system to
something it is not built to handle.