Also handle the later insertion programmatically
This might be a little bit of unnecessary complexity in the C++ code, but I wanted to make it simpler for our junior contributors to troubleshoot. By making the actual details of the string's position programmatic, they only need to worry about its relative position in the whole of responses. (A JSON comment was also added to accommodate this)

I have seen, and been told, and experienced that debugging the talk tests are quite troublesome! So hopefully this makes it a little easier.