units::temperature serialize and deserialize units
Serialize and deserialize units::temperature values as degrees celsius
offset from 0 degrees celsius. This means that temperature values are
saved with milli degrees celsius precision, which should not be plenty.

Add a round trip test to load/save/load units::temperature values and
ensure the in-memory values are not affected, and all the formats are
supported.

The legacy format "\"%f\"" needs to be supported because it exists in
save files as part of bodypart temperature calculations.