jmapgen_terrain: use string ids
Because conversion to an int_id requires that the object of that type is
already loaded, values loaded from JSON should be loaded as string ids,
then converted to int ids on use or after all JSON has been loaded.

Because jmapgen_terrain did not do this, if terrain definitions were
placed such that they were loaded after the palette the jmapgen_terrain
is part of, it would not be able to convert them to a valid int id.

This was triggered in the innawood mod, and was resolved by moving the
terrain defintions. Now that they are fixed, they can be restored to
their previous location.

TODO: memory profiling - what is the impact of this?