fix how selection changes are handled in a uilist with disabled entries
If you type HOME in a uilist and the first entry of the list happens
to be disabled then it keeps searching upwards in the list to find an
enabled entry to land on. But that makes it wrap around to the bottom
of the list, which is surprising. So don’t do that. But of course the
up arrow key should wrap around, so make sure to keep doing that. And
don’t forget that a similar thing could happen if you type END and end
up on a disabled item at the bottom of the list.

Fixes #83002