Move monster constructor initializations to header
Many monster class values are initialized in the header with different
values than are set in the constructor, after initializations were added
in c98bbcbc43b52635ba9a12fbee8bf29bb4236c1b.

For example, that initialized upgrade_time to 0, whereas it was set to
-1 in the constructor in 028cf26bdfda76345811f7aa017408f8871af611.

Because of the incorrect header value, the wrong default value was set
in 4813012dc17fdfeb8f238b40b40b59b06da3f049 when reading from a corpse
_without_ the variable already set.

Fix that bug, and move trivial initializations to the header.

Co-authored-by: fairyarmadillo <94415528+worm-girl@users.noreply.github.com>