add local minerals to mining item group
      { "item": "chunk_feldspar", "count": [ 10, 25 ], "prob": 50 },
      { "item": "chunk_graphite", "count": [ 0, 2 ], "prob": 2 },
      { "item": "chunk_caco3", "count": [ 0, 2 ], "prob": 60 }

rock walls are ambiguously either granite or limestone or marble
felspar makes up most of Massachusetts/NE granite
graphite and calcite crystals are present in the limestone/marble of Massachusetts, there are two graphite mines in the state and graphite crystals are present in the rock rarely. 
Calcite is extremely common in NE/Massachusetts. Fist sized crystals of Calcite are not hard to find