Replace the old pathfinding_settings with two new classes
PathfindingSettings is the direct replacement for the old
pathfinding_settings struct. It stores how a creature will react
to a given terrain, and what actions they can take.

RealityBubblePathfindingSettings contains settings that govern z-level
movement and parameters related to pathfinding, such as the maximum path
cost.

Currently the extra features of these two classes are completely unused,
as they are intended for integration with the new pathfinder.

Co-authored-by: prharvey <2677507+prharvey@users.noreply.github.com>