map boards display a map behind glass so just make them readable (#84253)
* map boards display a map behind glass so just make them readable

Currently you have to either pick the lock or smash the glass, then
pick up the map, then read it. But the whole point of a map board is
to display the map readably.

I thought about occasionally making the map board dirty, such as by
checking for bloodstains on the map tile, but that can be done later.

There currently can only be one examine action on a piece of furniture
so I ditched the lockpicking. Map boards that have previously been
picked will be migrated to smashed map boards. If you smash a map
board then you must pick the map up out of the wreckage before you can
read it.

Fixes #84205

* reformat json just to add noise to the commit history

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* clean up stuff left over from debugging

u_map_run_item_eocs already provides a message to the user in the case
of failure, although it’s not perfect for this situation. A second
message isn’t really useful.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>