Pick stuff up, put stuff down
Reverts test changes from https://github.com/CleverRaven/Cataclysm-DDA/pull/62723

...Those changes appear to test that the *loot sorting* option successfully unloads (?) items. As far as I can ask, that has never been the case.

I do not know how this test was passing before, but it was.

After some consultation with other devs on the devcord, we couldn't find a reason for it existing, so simply removing the erroneous fail was the decided upon course of action.