Better assignment of credit/blame for thrown molotovs
Also, have onfire effect credit/blame apply to the source of that effect

We were assigning `onfire` to bp_null as a default argument, which is all sorts of wrong. Fixed that by applying it to a random body part.

Lastly, forced the `onfire` effect to disappear on a Character once the affected limb is fully depleted. Instead of having a broken limb burning endlessly, to no effect.